third_party/blink/web_tests/external/wpt/mediacapture-region/CropTarget-fromElement.https.html

<!doctype html>
<html>

<head>
  <title>Test CropTarget.fromElement()</title>
  <meta name='assert' content='Test CropTarget.fromElement().' />
</head>

<body>
  <h1 class="instructions">Description</h1>
  <p class="instructions">
    This test checks for the behavior of <code>CropTarget.fromElement()</code>.
  </p>

  <div id='test-div'></div>
  <iframe id='test-iframe' src="about:blank"></iframe>
  <a id='test-a'></a>
  <div id='log'></div>

  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>

  <script>
    "use strict";

    promise_test(async () => {
      assert_true(!!CropTarget.fromElement);
      const crop_target = await CropTarget.fromElement(
        document.getElementById('test-iframe'));
      assert_equals(crop_target.constructor.name, 'CropTarget');
    }, "Produces a CropTarget for Elements of subtype iframe.");

    promise_test(async () => {
      assert_true(!!CropTarget.fromElement);
      const crop_target = await CropTarget.fromElement(
        document.getElementById('test-div'));
      assert_equals(crop_target.constructor.name, 'CropTarget');
    }, "Produces a CropTarget for Elements of subtype div.");

    promise_test(t => {
      assert_true(!!CropTarget.fromElement);
      return promise_rejects_js(t, TypeError,
        CropTarget.fromElement(undefined));
    }, "Rejects undefined with a TypeError.");

    promise_test(t => {
      assert_true(!!CropTarget.fromElement);
      return promise_rejects_js(t, TypeError, CropTarget.fromElement(123));
    }, "Rejects a non-Element with a TypeError.");

    promise_test(function (t) {
      assert_true(!!CropTarget.fromElement);

      return promise_rejects_dom(t, "NotSupportedError",
        CropTarget.fromElement(document.getElementById("test-a")));
    }, "Rejects unimplemented Element subtypes with a NotSupportedError");

    promise_test(async () => {
      assert_true(!!CropTarget.fromElement);

      const div_crop_target = await CropTarget.fromElement(
        document.getElementById('test-div'));
      assert_equals(div_crop_target.constructor.name, 'CropTarget');

      const iframe_crop_target = await CropTarget.fromElement(
        document.getElementById('test-iframe'));
      assert_equals(iframe_crop_target.constructor.name, 'CropTarget');

      assert_not_equals(div_crop_target, iframe_crop_target);
    }, "Distinct Elements produce distinct CropTargets.");

    promise_test(async () => {
      assert_true(!!CropTarget.fromElement);

      const div = document.getElementById('test-div');
      const div_crop_target = await CropTarget.fromElement(div);
      assert_equals(div_crop_target.constructor.name, 'CropTarget');

      const clone = div.cloneNode(true);
      document.querySelector('body').appendChild(clone);
      const clone_crop_target = await CropTarget.fromElement(clone);
      assert_equals(clone_crop_target.constructor.name, 'CropTarget');

      assert_not_equals(div_crop_target, clone_crop_target);
    }, "Cloned Elements produce distinct CropTargets.");

  </script>
</body>

</html>
